fixes #40 Fixes #41
Improve reconnect behaviour in local only lock service. Was silently trapping failures in certain cases and not reconnecting.
Bring in fixes to shared channel message handling where in some cases a message that was shutting down a virtual channel would shut down all of the virtual channels on the remote without closing the shared channel.
Since keep-alive behaviour is applied at the shared-channel level, this manifests as communication silently stopping in the higher level protocols without a failure recovery mechanism.

Fixed by not shutting down shared channels when a virtual channel message asks for it!